Addedmigrations and data seeting

+++ src/GamePacks.DataAccess/Models/Pack.cs
@@ -7,7 +7,7 @@ public class Pack
     public Guid Id {get;set;}
     public required string Name {get;set;}
     public required string ShortName {get;set;}
-    public long Price {get;set;}
+    public int Price {get;set;}
     public bool IsActive {get;set;}
     public Pack? ParentPack {get;set;}
     public ICollection<PackItem>? PackItems {get;set;}
